refactor: Unify path resolution and standardize on .todos.db
- Consolidate duplicate path resolution functions into single implementation
- Move primary implementation to datapath package to avoid circular imports
- Remove pkg/too/internal/helpers/helpers.go (duplicate code)
- Update all references from .todos.json to .todos.db
- Keep backward compatibility wrapper in pkg/too/path.go
