+++ pkg/too/commands/datapath/datapath.go
@@ -5,6 +5,53 @@ import (
 	"path/filepath"
 )
 
+// ResolveCollectionPath resolves the collection file path using the following order:
+// 1. If explicitPath is provided, use it as-is
+// 2. Search current directory and parent directories for .todos.db file (like git)
+// 3. Check TODO_DB_PATH environment variable
+// 4. Fall back to ~/.todos.db if it exists
+// 5. Default to .todos.db in current directory
+func ResolveCollectionPath(explicitPath string) string {
+	if explicitPath != "" {
+		return explicitPath
+	}
+
+	// Search upward for .todos.db file (like git does for .git)
+	dir, err := os.Getwd()
+	if err == nil {
+		for {
+			todosPath := filepath.Join(dir, ".todos.db")
+			if _, err := os.Stat(todosPath); err == nil {
+				return todosPath
+			}
+			
+			parent := filepath.Dir(dir)
+			if parent == dir {
+				// Reached root directory
+				break
+			}
+			dir = parent
+		}
+	}
+
+	// Check TODO_DB_PATH environment variable
+	if envPath := os.Getenv("TODO_DB_PATH"); envPath != "" {
+		return envPath
+	}
+
+	// Check if ~/.todos.db exists (home directory default)
+	home, err := os.UserHomeDir()
+	if err == nil {
+		homeDefault := filepath.Join(home, ".todos.db")
+		if _, err := os.Stat(homeDefault); err == nil {
+			return homeDefault
+		}
+	}
+
+	// Default to current directory
+	return ".todos.db"
+}
+
 // Options holds the options for the datapath command
 type Options struct {
 	CollectionPath string
@@ -17,58 +64,8 @@ type Result struct {
 
 // Execute shows the path to the data file
 func Execute(opts Options) (*Result, error) {
-	var storePath string
-
-	if opts.CollectionPath != "" {
-		// Use explicit path if provided
-		storePath = opts.CollectionPath
-	} else {
-		// Search upward for .todos file (like git does for .git)
-		dir, err := os.Getwd()
-		if err == nil {
-			found := false
-			for {
-				todosPath := filepath.Join(dir, ".todos.db")
-				if _, err := os.Stat(todosPath); err == nil {
-					storePath = todosPath
-					found = true
-					break
-				}
-				
-				parent := filepath.Dir(dir)
-				if parent == dir {
-					// Reached root directory
-					break
-				}
-				dir = parent
-			}
-			
-			if !found {
-				// Check TODO_DB_PATH environment variable
-				if envPath := os.Getenv("TODO_DB_PATH"); envPath != "" {
-					storePath = envPath
-				} else {
-					// Check if ~/.todos.db exists (home directory default)
-					home, err := os.UserHomeDir()
-					if err == nil {
-						homeDefault := filepath.Join(home, ".todos.db")
-						if _, err := os.Stat(homeDefault); err == nil {
-							storePath = homeDefault
-						} else {
-							// Default to current directory
-							storePath = ".todos.db"
-						}
-					} else {
-						// Fallback to current directory if can't get home
-						storePath = ".todos.db"
-					}
-				}
-			}
-		} else {
-			// Fallback to current directory if can't get working directory
-			storePath = ".todos.db"
-		}
-	}
+	// Use the unified path resolution function
+	storePath := ResolveCollectionPath(opts.CollectionPath)
 
 	// Get the absolute path
 	absPath, err := filepath.Abs(storePath)
